pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djangorestframework-mcp-server"
dynamic = ["version"]
description = "Expose djangorestframework-services services and selectors as an MCP (Model Context Protocol) server."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Artur Veres", email = "artur8118@gmail.com" },
]
keywords = ["django", "drf", "rest-framework", "mcp", "model-context-protocol", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "djangorestframework-services==0.6.0",
    "djangorestframework-dataclasses>=1.3.1",
    # ``typing_extensions.Unpack`` is used at runtime by the strict service /
    # selector Protocols (PEP 692) we re-export from the sister package.
    # Already pulled in transitively, but declared here so we don't depend on
    # an implicit contract.
    "typing_extensions>=4.6",
]

[project.optional-dependencies]
toon = ["python-toon>=0.1"]
oauth = ["django-oauth-toolkit>=2.3"]
redis = ["redis>=5.0"]
otel = ["opentelemetry-api>=1.27"]

[project.urls]
Homepage = "https://github.com/Artui/djangorestframework-mcp-server"
Repository = "https://github.com/Artui/djangorestframework-mcp-server"
Issues = "https://github.com/Artui/djangorestframework-mcp-server/issues"

[tool.hatch.version]
path = "rest_framework_mcp/version.py"
pattern = "^__version__(?:\\s*:\\s*[A-Za-z][A-Za-z0-9_]*)?\\s*=\\s*[\"'](?P<version>[^\"']+)[\"']"

[tool.hatch.build.targets.wheel]
packages = ["rest_framework_mcp"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "ty>=0.0.32",
    "pre-commit>=3.8.0",
    "pytest>=8.0",
    "pytest-django>=4.11.1",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "flake8-tidy-imports>=4.12.0",
    "django-oauth-toolkit>=2.3",
    "django-stubs>=5.0",
    "djangorestframework-stubs>=3.15",
    "redis>=5.0",
    "fakeredis>=2.20",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.conftest_settings"
asyncio_mode = "auto"
addopts = "--cov=rest_framework_mcp --cov-report=term-missing --cov-fail-under=100"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["rest_framework_mcp"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I", "TID"]
ignore = ["E203", "E501", "E231", "F811", "E722", "E741"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["rest_framework_mcp"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
root = ["./rest_framework_mcp"]

# bump-my-version drives `make release-bump` — it rewrites the version in
# rest_framework_mcp/version.py and promotes the CHANGELOG [Unreleased]
# block under a new dated heading. Run via `uvx bump-my-version` (no
# permanent dependency added to the project).
[tool.bumpversion]
current_version = "0.1.0"
allow_dirty = false
commit = false
tag = false

[[tool.bumpversion.files]]
filename = "rest_framework_mcp/version.py"
search = '__version__: str = "{current_version}"'
replace = '__version__: str = "{new_version}"'

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "## [Unreleased]"
replace = """## [Unreleased]

## [{new_version}] — {now:%Y-%m-%d}"""

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "[Unreleased]: https://github.com/Artui/djangorestframework-mcp-server/compare/v{current_version}...HEAD"
replace = """[Unreleased]: https://github.com/Artui/djangorestframework-mcp-server/compare/v{new_version}...HEAD
[{new_version}]: https://github.com/Artui/djangorestframework-mcp-server/compare/v{current_version}...v{new_version}"""
